stylized formReg and buttons
