+++ app/[locale]/(auth)/signup/page.tsx
@@ -1,0 +1,16 @@
+'use client'
+
+import { SignUpForm } from '@/components/SignUpForm/SignUpForm'
+
+const SignupPage = () => {
+    return (
+        <div>
+            <h1 className='container mx-auto flex items-center flex-col'>
+                Registration
+            </h1>
+            <SignUpForm />
+        </div>
+    )
+}
+
+export default SignupPage
