+++ components/Header/Header.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link'
 import LocaleSwitcher from '../LocaleSwitcher/LocaleSwitcher'
 import { useTranslations } from 'next-intl'
 import SignInBtn from '../SignInBtn/SignInBtn'
-import SignInButton from '../SignInButton'
+import SignInButton from '../SignInBtn/SignInButton'
 
 export default function Header() {
     const t = useTranslations('Navigation')
@@ -18,6 +18,7 @@ export default function Header() {
             </Link>
 
             <Navigation />
+            {/* <SignInBtn navLinks={[]} /> */}
             <SignInButton />
             <LocaleSwitcher />
         </div>
