+++ components/UI/Input/Input.scss
@@ -1,4 +1,4 @@
-form {
+.form {
     display: flex;
     flex-direction: column;
     justify-content: center;
@@ -14,6 +14,33 @@ form {
         flex-direction: column;
         gap: 5px;
         width: -webkit-fill-available;
+    }
+
+    .input input {
+        background: none;
+        border: 2px solid #393d40;
+        padding: 10px 15px;
+        border-radius: 4px;
+        color: #c9cccf;
+        font-size: 15px;
+    }
+}
+
+.form-register {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    width: min(650px, 100%);
+    margin: auto;
+    gap: 30px;
+    border: 1px solid #393d40;
+    padding: 20px;
+    border-radius: 0.75rem;
+    .input {
+        display: flex;
+        flex-direction: column;
+        gap: 5px;
+        // width: -webkit-fill-available;
     }
 
     .input input {
