+++ styles/globals.css
@@ -34,7 +34,7 @@ h1 {
     color: #c9cccf;
     border-radius: 15px;
     padding: 5px 10px;
-    margin: 0px 10px;
+    /* margin: 0px 10px; */
     font-family:
         system-ui,
         -apple-system,
@@ -53,4 +53,30 @@ h1 {
 .button:hover {
     background-color: #c9cccf;
     color: #232324;
+}
+
+.button-signUp {
+    background-color: mediumseagreen;
+    color: white;
+    border-radius: 15px;
+    padding: 5px 10px;
+    /* margin: 0px 10px; */
+    font-family:
+        system-ui,
+        -apple-system,
+        BlinkMacSystemFont,
+        'Segoe UI',
+        Roboto,
+        Oxygen,
+        Ubuntu,
+        Cantarell,
+        'Open Sans',
+        'Helvetica Neue',
+        sans-serif;
+    cursor: pointer;
+}
+
+.button-signUp:hover {
+    background-color: rgb(138, 207, 169);
+    /* color: #232324; */
 }
